Rename App -> Workplace in connector files
